+++ src/commands/sbom/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod device;
 pub mod generate;